Move some files about.
