+++ tests/test_inputs.py
@@ -3,7 +3,7 @@
 # pylint: disable=protected-access,no-self-use
 from unittest import TestCase
 
-from inputs.devices.event import InputEvent
+from inputs.platforms.event import InputEvent
 from inputs.platforms.c import convert_timeval
 from inputs.utils import get_key, get_mouse, get_gamepad
 from inputs.errors import UnpluggedError
